Move input_common.h to the input service folder
