--- lib/input_common.scons
@@ -1,7 +1,0 @@
-Import("env")
-
-env.AppendUnique(
-    CPPPATH=[
-        Dir("input_common").srcnode(),
-    ],
-)
