+++ lib/register_bsb_common_libs.scons
@@ -22,7 +22,6 @@ for libname in [
     "flipper_application",
     "nanopb",
     "heatshrink",
-    "input_common",
     "ext_applications",
     "firmware_applications_f20",
     ## SI917
