+++ targets/f20/target.json
@@ -116,7 +116,6 @@
         "furi",
         "drivers",
         "version",
-        "input_common",
         "ext_applications",
         "firmware_applications_f20"
     ],
