+++ targets/f64/target.json
@@ -82,7 +82,6 @@
         "targets",
         "furi",
         "toolbox_f64",
-        "input_common",
         "firmware_applications_f64"
     ],
     "fw_modules": [
